fix: PixelOp::apply takes &mut self for RowConverter scratch buffers
RowConverter::convert_row requires &mut self to reuse internal scratch
buffers. Updated PixelOp trait and all impls/call sites accordingly.
Also removed unnecessary `as usize` cast flagged by clippy.

+++ src/graph.rs
@@ -1289,7 +1289,7 @@ fn detect_content_bounds(
         return (0, 0, w, h);
     }
 
-    let bpp = mat.format().bytes_per_pixel() as usize;
+    let bpp = mat.format().bytes_per_pixel();
 
     // Reference color: top-left pixel
     let row0 = mat.row(0);
